creates Smart class
