+++ src/view/smart.js
@@ -1,0 +1,44 @@
+import Abstract from './abstract.js';
+
+export default class Smart extends Abstract {
+  constructor() {
+    super();
+    this._state = {};
+  }
+
+  updateState(update, justStateUpdate) {
+    if (!update) {
+      return;
+    }
+
+    this._state = Object.assign(
+      {},
+      this._state,
+      update,
+    );
+
+    if (justStateUpdate) {
+      return;
+    }
+
+    this.updateElement();
+  }
+
+  updateElement() {
+    const prevElement = this.getElement();
+    const parent = prevElement.parentElement;
+    const prevScrollPosition = prevElement.scrollTop;
+    this.removeElement();
+
+    const newElement = this.getElement();
+
+    parent.replaceChild(newElement, prevElement);
+    newElement.scrollTop = prevScrollPosition;
+
+    this.restoreHandlers();
+  }
+
+  restoreHandlers() {
+    throw new Error('Abstract method not implemented: restoreHandlers');
+  }
+}
